rename loginSlice -> userSlice
