+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Outlet } from 'react-router-dom';
 import './App.css';
 import NavBar from './components/NavBar';
 import { useAppSelector } from './redux/hooks';
-import { selectIsToggleMenu } from './redux/loginSlice';
+import { selectIsToggleMenu } from './redux/userSlice';
 
 function App() {
   const toggled = useAppSelector(selectIsToggleMenu);
