+++ src/pages/Editor/Editor.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { v4 as uuidv4 } from 'uuid';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
-import { addNote, editNote, selectProfile } from '../../redux/loginSlice';
 import { NoteType } from '../../redux/types';
+import { addNote, editNote, selectProfile } from '../../redux/userSlice';
 import * as styles from '../../utils/styles';
 import CommentBox from './CommentBox';
 import EditorHelper from './EditorHelper';
