+++ src/pages/Landing/Home.tsx
@@ -3,8 +3,8 @@ import { Link, useSearchParams } from 'react-router-dom';
 import Swal from 'sweetalert2';
 import Header from '../../components/header';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
-import { addNote, selectProfile } from '../../redux/loginSlice';
 import { NoteType } from '../../redux/types';
+import { addNote, selectProfile } from '../../redux/userSlice';
 import { firestore } from '../../utils/firebase';
 import * as styles from '../../utils/styles';
 import { parseDate, parseWeekday } from '../../utils/utils';
