+++ src/redux/store.ts
@@ -1,9 +1,9 @@
 import { Action, ThunkAction, configureStore } from '@reduxjs/toolkit';
-import loginReducer from './loginSlice';
+import userReducer from './userSlice';
 
 export const store = configureStore({
   reducer:{
-    login: loginReducer
+    user: userReducer
   }
 });
